make stonex mock fixture
