+++ stonex/tests/test_account.cpp
@@ -1,52 +1,20 @@
 #include "stonex.h"
+#include "stonex_mock.hpp"
 #include "trading_types.hpp"
 #include <cpr/cpr.h>
 #include <ctime>
-#include <gmock/gmock.h>
 #include <gtest/gtest.h>
-#include <httplib.h>
 #include <iostream>
 #include <nlohmann/json.hpp>
 
 using json = nlohmann::json;
-
-class StoneXRespTest : public ::testing::Test {
-protected:
-  httplib::Server svr;
-  std::thread server_thread;
-
-  void SetUp() override {
-    setenv("STONEX_APPKEY", "TESTING", 1);
-    setenv("STONEX_USER", "TESTING", 1);
-    setenv("STONEX_PASS", "TESTING", 1);
-    StoneX::base_url_ptr->assign("http://localhost:8080");
-    StoneX::trading_url_ptr->assign("http://localhost:8080");
-
-    // Bind the server before running it
-    svr.bind_to_port("localhost", 8080);
-
-    // Start the server in a separate thread
-    server_thread = std::thread([this]() { svr.listen_after_bind(); });
-
-    // Give the server time to start
-    std::this_thread::sleep_for(std::chrono::milliseconds(200));
-  }
-
-  void TearDown() override {
-    // Stop the server and join the thread
-    svr.stop();
-    if (server_thread.joinable()) {
-      server_thread.join();
-    }
-  }
-};
 
 TEST_F(StoneXRespTest, TestInitializeSession) {
 
   StoneX::ApiLogOnResponseDTO sess_resp;
   sess_resp.session = "1234567890";
   // Configure mock responses
-  svr.Post("/Session", [&](const httplib::Request &req, httplib::Response &res) {
+  svr->Post("/Session", [&](const httplib::Request &req, httplib::Response &res) {
     res.set_content(json(sess_resp).dump(), "application/json");
   });
 
@@ -54,10 +22,10 @@ TEST_F(StoneXRespTest, TestInitializeSession) {
   account_resp.tradingAccounts.push_back(StoneX::ApiTradingAccountDTO{
       1234, 456, "test_code", "test_status", "test_type", "test_name"});
 
-  svr.Get("/UserAccount/ClientAndTradingAccount",
-          [account_resp](const httplib::Request &req, httplib::Response &res) {
-            res.set_content(json(account_resp).dump(), "application/json");
-          });
+  svr->Get("/UserAccount/ClientAndTradingAccount",
+           [account_resp](const httplib::Request &req, httplib::Response &res) {
+             res.set_content(json(account_resp).dump(), "application/json");
+           });
 
   EXPECT_EQ(StoneX::AccountCredentials::initialized, false);
